CONSOLE-4858: Resolve min-length cycles
