+++ frontend/packages/console-shared/src/components/namespace/index.ts
@@ -1,2 +1,2 @@
 export { isCurrentUser, isSystemNamespace, isOtherUser, requesterFilter } from './filters';
-export { default as NamespaceDropdown } from './NamespaceDropdown';
+export { NamespaceDropdown } from './NamespaceDropdown';
