+++ frontend/packages/console-shared/src/hooks/useNotificationAlerts.ts
@@ -2,7 +2,7 @@ import { useEffect, useMemo, useState } from 'react';
 import * as _ from 'lodash';
 import { useSelector } from 'react-redux';
 import { HIDE_USER_WORKLOAD_NOTIFICATIONS_USER_SETTINGS_KEY } from '@console/app/src/consts';
-import { useNamespacedNotificationAlertsPoller } from '@console/app/src/hooks/useNotificationPoller';
+import { useNamespacedNotificationAlertsPoller } from '@console/app/src/hooks/useNamespacedNotificationAlertsPoller';
 import { LabelSelector, ObjectMetadata } from '@console/internal/module/k8s';
 import { NotificationAlerts } from '@console/internal/reducers/observe';
 import { RootState } from '@console/internal/redux';
